requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
psutil>=5.9.0
cryptography>=3.4.8
urllib3>=1.26.0

